requirements.txt
streamlit
requests
beautifulsoup4
googlesearch-python
pandas
apscheduler
